Introduce module `DatabaseLimits`
To prepare longer identifier support in nextval branch.

Quoting module has implicit length information for
`NONQUOTED_OBJECT_NAME` and `NONQUOTED_DATABASE_LINK`, which will be
addressed in another commit.
